Input: Allow mapping buttons/shortcuts to mouse buttons + refactoring (16-bit keycodes, make mouse buttons behave like all other keys)

+++ Core/Shared/BaseControlDevice.cpp
@@ -228,7 +228,7 @@ void BaseControlDevice::InvertBit(uint8_t bit)
 	}
 }
 
-void BaseControlDevice::SetPressedState(uint8_t bit, uint32_t keyCode)
+void BaseControlDevice::SetPressedState(uint8_t bit, uint16_t keyCode)
 {
 	if(KeyManager::IsKeyPressed(keyCode)) {
 		SetBit(bit);
